Assignments/Week-13/requirements.txt
flask
python-dotenv
openai
azure-core
